+++ packages/prop-house-webapp/src/hooks/useContentModeration.ts
@@ -8,71 +8,6 @@ const supabase = createClient(
   process.env.REACT_APP_SUPABASE_KEY!,
 );
 
-export const useHiddenPropIds = (roundAddress: string) => {
-  const [hiddenPropIds, setHiddenPropIds] = useState<number[] | undefined>();
-  const [shouldFetch, setShouldFetch] = useState(true);
-
-  useEffect(() => {
-    if (!shouldFetch) return;
-
-    const fetchHiddenPropIds = async (roundAddress: string) => {
-      setShouldFetch(false);
-      try {
-        const { data: proposals } = await supabase
-          .from('proposals')
-          .select()
-          .eq('roundAddress', roundAddress);
-
-        if (!proposals) return;
-
-        const propIds = proposals.map(p => p.proposalId);
-        setHiddenPropIds(propIds);
-      } catch (e) {
-        console.log(e);
-      }
-    };
-
-    fetchHiddenPropIds(roundAddress);
-  }, [roundAddress, shouldFetch]);
-
-  const refresh = () => {
-    setShouldFetch(true);
-  };
-
-  return { hiddenPropIds, refresh };
-};
-
-export const useIsHiddenRound = (roundAddress: string) => {
-  const [isHidden, setIsHidden] = useState<boolean | undefined>();
-  const [shouldFetch, setShouldFetch] = useState(true);
-
-  useEffect(() => {
-    if (!shouldFetch) return;
-
-    const fetchIsHidden = async (roundAddress: string) => {
-      setShouldFetch(false);
-      try {
-        const { data: round } = await supabase
-          .from('rounds')
-          .select()
-          .eq('address', roundAddress)
-          .eq('hidden', true)
-          .single();
-
-        setIsHidden(round ? true : false);
-      } catch (e) {
-        console.log(e);
-      }
-    };
-    fetchIsHidden(roundAddress);
-  }, [shouldFetch, roundAddress]);
-
-  const refresh = () => {
-    setShouldFetch(true);
-  };
-  return { isHidden, refresh };
-};
-
 /**
  * Supplying a house will check for mod status as a house admin or superadmin.
  * Not supplying it will only check for superadmin status.
